% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_graph.R
\name{select_by_prefix}
\alias{select_by_prefix}
\title{Filter a graph by discipline}
\usage{
select_by_prefix(g, prefixes)
}
\arguments{
\item{g}{A directed igraph or tidygraph course graph with unique course_code attributes.}

\item{prefixes}{One or more four-letter uppercase discipline prefixes.}
}
\value{
A graph, which may be empty if no prefix matches.
}
\description{
Removes nonmatching nodes, potentially removing intermediate paths. Use plot highlighting to retain context.
}
